preprocessing/requirements.txt
huggingface_hub==0.36.2
numpy==2.2.6
opencv_python_headless==4.13.0.92
pandas==2.3.3
Pillow==12.0.0
torch==2.10.0+cu126
torchvision==0.25.0+cu126
transformers==4.36.2
open_clip_torch==3.2.0
openpyxl==3.1.5